backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
scanpy==1.10.3
matplotlib==3.9.0
numpy==2.0.0
pandas==2.2.0
python-multipart==0.0.17
python-dotenv==1.0.0
